+++ pom.xml
@@ -1,26 +1,27 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  Copyright 2014 EsmerilProgramming.
+  JBoss, Home of Professional Open Source Copyright 2014 Red Hat Inc. and/or
+  its affiliates and other contributors as indicated by the @authors tag. 
+  All rights reserved.
+  See the copyright.txt
+  in the distribution for a full listing of individual contributors.
 
   Licensed under the Apache License, Version 2.0 (the "License"); you may
   not use this file except
   in compliance with the License. You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software distributed
-  under the License
-  is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
-  ANY KIND, either express
-  or implied. See the License for the specific language governing permissions
-  and limitations under
-  the License.
+  http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable
+  law or agreed to in
+  writing, software distributed under the License is distributed on an "AS
+  IS" BASIS, WITHOUT
+  WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+  License for the specific
+  language governing permissions and limitations under the License.
 -->
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>org.esmerilprogramming</groupId>
+  <groupId>org.jboss.aesh</groupId>
   <artifactId>mterm</artifactId>
   <version>0.56</version>
 
@@ -60,11 +61,11 @@
         </executions>
         <configuration>
           <executable>java</executable>
-          <mainClass>org.esmerilprogramming.mterm.Mterm</mainClass>
+          <mainClass>org.jboss.aesh.mterm.Mterm</mainClass>
         </configuration>
       </plugin>
 
-     <!-- this put all dependencies on lib directory to distribution -->
+      <!-- this put all dependencies on lib directory to distribution -->
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-dependency-plugin</artifactId>
@@ -95,7 +96,7 @@
             <manifest>
               <addClasspath>true</addClasspath>
               <classpathPrefix>lib/</classpathPrefix>
-              <mainClass>org.esmerilprogramming.mterm.Mterm</mainClass>
+              <mainClass>org.jboss.aesh.mterm.Mterm</mainClass>
             </manifest>
             <manifestEntries>
               <Class-Path>.</Class-Path>
