--- src/main/java/org/esmerilprogramming/mterm/gui/Menu.java
@@ -1,60 +1,0 @@
-/*
- * Copyright 2014 EsmerilProgramming.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- * 
- * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-package org.esmerilprogramming.mterm.gui;
-
-import javax.swing.ImageIcon;
-import javax.swing.JMenu;
-import javax.swing.JMenuBar;
-import javax.swing.JMenuItem;
-
-import org.esmerilprogramming.mterm.action.menu.MenuBaseAction;
-
-/**
- * The menu class.
- *
- * @author <a href="mailto:00hf11@gmail.com">Helio Frota</a>
- */
-public class Menu {
-
-  private JMenuBar menuBar;
-
-  public Menu() {
-    menuBar = new JMenuBar();
-  }
-
-  public Menu addMenu(String m) {
-    menuBar.add(new JMenu(m));
-    return this;
-  }
-
-  public Menu addSubMenu(int index, String sm) {
-    menuBar.getMenu(index).add(new JMenuItem(sm));
-    return this;
-  }
-
-  public Menu addSubMenu(int index, String sm, ImageIcon icon) {
-    menuBar.getMenu(index).add(new JMenuItem(new MenuBaseAction(sm, icon)));
-    return this;
-  }
-
-  public Menu addSubMenu(int index, MenuBaseAction menuAction) {
-    menuBar.getMenu(index).add(new JMenuItem(menuAction));
-    return this;
-  }
-
-  public JMenuBar create() {
-    return this.menuBar;
-  }
-
-}
