+++ src/main/java/org/jboss/aesh/mterm/action/menu/MenuAboutAction.java
@@ -1,0 +1,49 @@
+/*
+ * JBoss, Home of Professional Open Source Copyright 2014 Red Hat Inc. and/or its affiliates and
+ * other contributors as indicated by the @authors tag. All rights reserved. See the copyright.txt
+ * in the distribution for a full listing of individual contributors.
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law or agreed to in
+ * writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
+ * language governing permissions and limitations under the License.
+ */
+package org.jboss.aesh.mterm.action.menu;
+
+import java.awt.event.ActionEvent;
+
+import javax.swing.ImageIcon;
+
+import org.jboss.aesh.mterm.gui.MessageDialog;
+
+/**
+ * Menu MenuAboutAction class.
+ *
+ * @author <a href="mailto:00hf11@gmail.com">Helio Frota</a>
+ */
+@SuppressWarnings("serial")
+public class MenuAboutAction extends MenuBaseAction {
+
+    public MenuAboutAction(String text, ImageIcon icon) {
+        super(text, icon);
+    }
+
+    public void actionPerformed(ActionEvent e) {
+        StringBuilder sb = new StringBuilder();
+        for (int i = 0; i < 20; i++) {
+            sb.append(" ");
+        }
+        sb.append("\n-=-=-=-= Mterm =-=-=-=-");
+        sb.append("\n\n\nA simple java terminal emulator");
+        for (int i = 0; i < 10; i++) {
+            sb.append(" ");
+        }
+        for (int i = 0; i < 5; i++) {
+            sb.append("\n");
+        }
+        sb.append("Created by Aesh Team");
+        new MessageDialog().info(sb.toString());
+    }
+}
