+++ src/main/java/org/jboss/aesh/mterm/action/menu/MenuNewAction.java
@@ -1,0 +1,42 @@
+/*
+ * JBoss, Home of Professional Open Source Copyright 2014 Red Hat Inc. and/or its affiliates and
+ * other contributors as indicated by the @authors tag. All rights reserved. See the copyright.txt
+ * in the distribution for a full listing of individual contributors.
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law or agreed to in
+ * writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
+ * language governing permissions and limitations under the License.
+ */
+package org.jboss.aesh.mterm.action.menu;
+
+import java.awt.event.ActionEvent;
+
+import javax.swing.ImageIcon;
+import javax.swing.SwingUtilities;
+
+import org.jboss.aesh.mterm.gui.MtermUI;
+
+/**
+ * Menu MenuNewAction class.
+ *
+ * @author <a href="mailto:00hf11@gmail.com">Helio Frota</a>
+ */
+@SuppressWarnings("serial")
+public class MenuNewAction extends MenuBaseAction {
+
+    public MenuNewAction(String text, ImageIcon icon) {
+        super(text, icon);
+    }
+
+    public void actionPerformed(ActionEvent e) {
+        SwingUtilities.invokeLater(new Runnable() {
+            @Override
+            public void run() {
+                new MtermUI();
+            }
+        });
+    }
+}
