+++ src/main/resources/org/jboss/aesh/mterm/gui/icons/about.png
