+++ src/main/resources/org/jboss/aesh/mterm/gui/icons/full_screen.png
