+++ src/main/resources/org/jboss/aesh/mterm/gui/icons/new.png
