+++ src/main/resources/org/jboss/aesh/mterm/gui/icons/set-title.png
